+++ src/ticket/ticketventa.java
@@ -14,7 +14,7 @@ public class ticketventa {
                ArrayList piezas, 
                ArrayList precio, 
                ArrayList importe, 
-               float subtotal, float total, float pago, float cambio, int numerodeventa) {
+               float subtotal, float total, float pago, float cambio, long numerodeventa) {
        
         Date date = new Date();
         DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
